stockkkkk/requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1

# ── Data & Finance ─────────────────────────────────────────────────────────────
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0

# ── Machine Learning ───────────────────────────────────────────────────────────
scikit-learn>=1.3.0
tensorflow>=2.13.0      # includes Keras; omit if you want CPU-only: tensorflow-cpu

# ── HTTP & Utilities ───────────────────────────────────────────────────────────
httpx>=0.27.0
requests>=2.31.0
